requirements.txt
bleach
genshi
jinja2
peewee
pillow
pyyaml
tweepy
requests
selenium
feedparser
readability-lxml
git+https://github.com/edsu/htmldiff#egg=htmldiff
